Add missing tests for ViewModels

+++ core/src/main/java/com/marknjunge/core/data/model/OrderItem.kt
@@ -16,7 +16,7 @@ data class OrderItem(
 	val totalPrice: Double,
 
 	@SerialName("productBasePrice")
-	val productBasePrice: Int,
+	val productBasePrice: Double,
 
 	@SerialName("productName")
 	val productName: String,
